use placholder class in custom job spec
