+++ sdk/python/kfp/dsl/_container_op.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import copy
 import inspect
 import re
 import warnings
@@ -26,6 +27,7 @@ from kubernetes.client.models import (V1Container, V1EnvVar, V1EnvFromSource,
 
 from kfp.components import _structures
 from kfp.dsl import _pipeline_param
+from kfp.dsl import dsl_utils
 from kfp.pipeline_spec import pipeline_spec_pb2
 
 # generics
@@ -1369,15 +1371,14 @@ class ContainerOp(BaseOp):
     super(ContainerOp, self).add_node_selector_constraint(label_name, value)
     return self
 
-  def set_custom_job_spec(self,
-                          custom_job_spec: Dict[str, Any]) -> 'ContainerOp':
+  def set_custom_job_spec(self, job_spec: Dict[str, Any]) -> 'ContainerOp':
     """Sets custom job spec.
 
     When compiling for v2, this function can be used to set custom job spec used
     for AI Platform (Unified) service.
 
     Args:
-      custom_job_spec: JSON struct of the CustomJob spec, representing the job
+      job_spec: JSON struct of the CustomJob spec, representing the job
         that will be submitted to AI Platform (Unified) service. See
         https://cloud.google.com/ai-platform-unified/docs/reference/rest/v1beta1/CustomJobSpec
         for detailed reference.
@@ -1385,26 +1386,45 @@ class ContainerOp(BaseOp):
     Returns:
       self return to allow chained call with other set method.
     """
-    self.custom_job_spec = custom_job_spec
-    if 'jobSpec' not in self.custom_job_spec or \
-       'workerPoolSpecs' not in self.custom_job_spec['jobSpec']:
-      worker_pool_spec = {
+    job_spec = copy.deepcopy(job_spec)
+
+    def _is_output_parameter(output_key: str) -> bool:
+      return output_key in (
+          self.component_spec.output_definitions.parameters.keys())
+
+    if 'workerPoolSpecs' not in job_spec:
+      job_spec['workerPoolSpecs'] = [{
           'machineSpec': {
               'machineType': _DEFAULT_CUSTOM_JOB_MACHINE_TYPE
           },
           'replicaCount': '1',
-          'containerSpec': {
-              'imageUri': self._container.image,
-          }
-      }
-      if self._container.command:
-        worker_pool_spec['containerSpec']['command'] = self._container.command
-      if self._container.args:
-        worker_pool_spec['containerSpec']['args'] = self._container.args
+      }]
+    for worker_pool_spec in job_spec['workerPoolSpecs']:
+      if 'containerSpec' in worker_pool_spec:
+        container_spec = worker_pool_spec['containerSpec']
+        if 'command' in container_spec:
+          dsl_utils.resolve_cmd_lines(container_spec['command'], _is_output_parameter)
+        if 'args' in container_spec:
+          dsl_utils.resolve_cmd_lines(container_spec['args'], _is_output_parameter)
 
-      if 'jobSpec' not in self.custom_job_spec:
-        self.custom_job_spec['jobSpec'] = {}
-      self.custom_job_spec['jobSpec']['workerPoolSpecs'] = [worker_pool_spec]
+      elif 'pythonPackageSpec' in worker_pool_spec:
+        # For custom Python training, resolve placeholders in args only.
+        python_spec = worker_pool_spec['pythonPackageSpec']
+        if 'args' in python_spec:
+          dsl_utils.resolve_cmd_lines(python_spec['args'], _is_output_parameter)
+
+      else:
+        # If neither 'containerSpec' nor 'pythonPackageSpec' presents, fill in
+        # 'containerSpec' using existing information.
+        worker_pool_spec['containerSpec'] = {
+            'imageUri': self._container.image,
+        }
+        if self._container.command:
+          worker_pool_spec['containerSpec']['command'] = self._container.command
+        if self._container.args:
+          worker_pool_spec['containerSpec']['args'] = self._container.args
+
+    self.custom_job_spec = {'name': self.name, 'jobSpec': job_spec}
 
     return self
