+++ sdk/python/kfp/dsl/dsl_utils.py
@@ -14,12 +14,23 @@
 """Utilities functions KFP DSL."""
 
 import re
-from typing import Union
+from typing import Callable, List, Optional, Union
+
+from kfp.components import _structures
 from kfp.pipeline_spec import pipeline_spec_pb2
 
 _COMPONENT_NAME_PREFIX = 'comp-'
 _TASK_NAME_PREFIX = 'task-'
 _EXECUTOR_LABEL_PREFIX = 'exec-'
+
+# TODO: Support all declared types in
+# components._structures.CommandlineArgumenType
+_CommandlineArgumentType = Union[str, int, float,
+                                 _structures.InputValuePlaceholder,
+                                 _structures.InputPathPlaceholder,
+                                 _structures.OutputPathPlaceholder,
+                                 _structures.InputUriPlaceholder,
+                                 _structures.OutputUriPlaceholder,]
 
 
 def sanitize_component_name(name: str) -> str:
@@ -77,3 +88,60 @@ def remove_task_name_prefix(sanitized_task_name: str) -> str:
   """
   assert sanitized_task_name.startswith(_TASK_NAME_PREFIX)
   return sanitized_task_name[len(_TASK_NAME_PREFIX):]
+
+
+# TODO: share with dsl._component_bridge
+def _input_artifact_uri_placeholder(input_key: str) -> str:
+  return "{{{{$.inputs.artifacts['{}'].uri}}}}".format(input_key)
+
+
+def _input_artifact_path_placeholder(input_key: str) -> str:
+  return "{{{{$.inputs.artifacts['{}'].path}}}}".format(input_key)
+
+
+def _input_parameter_placeholder(input_key: str) -> str:
+  return "{{{{$.inputs.parameters['{}']}}}}".format(input_key)
+
+
+def _output_artifact_uri_placeholder(output_key: str) -> str:
+  return "{{{{$.outputs.artifacts['{}'].uri}}}}".format(output_key)
+
+
+def _output_artifact_path_placeholder(output_key: str) -> str:
+  return "{{{{$.outputs.artifacts['{}'].path}}}}".format(output_key)
+
+
+def _output_parameter_path_placeholder(output_key: str) -> str:
+  return "{{{{$.outputs.parameters['{}'].output_file}}}}".format(output_key)
+
+
+def resolve_cmd_lines(cmds: Optional[List[_CommandlineArgumentType]],
+                       is_output_parameter: Callable[[str], bool]) -> None:
+  """Resolves a list of commands/args."""
+
+  def _resolve_cmd(cmd: Optional[_CommandlineArgumentType]) -> Optional[str]:
+    """Resolves a single command line cmd/arg."""
+    if cmd is None:
+      return None
+    elif isinstance(cmd, (str, float, int)):
+      return str(cmd)
+    elif isinstance(cmd, _structures.InputValuePlaceholder):
+      return _input_parameter_placeholder(cmd.input_name)
+    elif isinstance(cmd, _structures.InputPathPlaceholder):
+      return _input_artifact_path_placeholder(cmd.input_name)
+    elif isinstance(cmd, _structures.InputUriPlaceholder):
+      return _input_artifact_uri_placeholder(cmd.input_name)
+    elif isinstance(cmd, _structures.OutputPathPlaceholder):
+      if is_output_parameter(cmd.output_name):
+        return _output_parameter_path_placeholder(cmd.output_name)
+      else:
+        return _output_artifact_path_placeholder(cmd.output_name)
+    elif isinstance(cmd, _structures.OutputUriPlaceholder):
+      return _output_artifact_uri_placeholder(cmd.output_name)
+    else:
+      raise TypeError('Got unexpected placeholder type for %s' % cmd)
+
+  if not cmds:
+    return
+  for idx, cmd in enumerate(cmds):
+    cmds[idx] = _resolve_cmd(cmd)
